requirements.txt
appdirs==1.4.3
astroid==1.5.2
decorator==4.0.11
Django==1.11
django-braces==1.11.0
django-cors-headers==2.0.2
django-extensions==1.7.8
django-filter==1.0.2
django-oauth-toolkit==0.12.0
djangorestframework==3.6.2
flake8==3.3.0
ipython==5.3.0
ipython-genutils==0.2.0
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.8
mccabe==0.6.1
oauthlib==2.0.1
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.10.10
